hotfix 20.2.7

+++ malloc_free/2-str_concat.c
@@ -10,13 +10,13 @@
 
 char *str_concat(char *s1, char *s2)
 {
-	int i, j, firstlength, secondlength;
+	unsigned int i, j, firstlength, secondlength;
 	char *newstring;
 
 	if (s1 != NULL)
-	firstlength = (int)strlen(s1);
+	firstlength = strlen(s1);
 	if (s2 != NULL)
-	secondlength = (int)strlen(s2);
+	secondlength = strlen(s2);
 
 	newstring = malloc(sizeof(char) * (firstlength + secondlength + 1));
 	if (newstring == NULL)
